+++ src/shared/lib/html/table-validation.js
