+++ src/shared/scrapers/USA/PA/index.js
@@ -2,7 +2,7 @@ import * as fetch from '../../../lib/fetch/index.js';
 import * as parse from '../../../lib/parse.js';
 import * as transform from '../../../lib/transform.js';
 import * as geography from '../../../lib/geography/index.js';
-import * as htmlTableValidation from '../../../lib/html-table-validator.js';
+import * as htmlTableValidation from '../../../lib/html/table-validation.js';
 
 // Set county to this if you only have state data, but this isn't the entire state
 // const UNASSIGNED = '(unassigned)';
